improve metrics
