+++ requirements.txt
@@ -3,3 +3,4 @@ flasgger==0.9.7.1
 flask-cors==5.0.1
 requests==2.32.2
 lib_version @ git+https://github.com/remla25-team13/lib-version.git
+prometheus-flask-exporter==0.23.2
